trim down component wrappers
